telemetry_service/requirements.txt
# telemetry_service/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.0
pydantic-settings>=2.0
sqlalchemy[asyncio]>=2.0
asyncpg>=0.29.0
celery[redis]>=5.3.0
redis>=5.0.0
sentry-sdk[fastapi]>=1.39.0
google-cloud-logging>=3.8.0
google-cloud-bigquery>=3.14.0
numpy>=1.24.0
pytest>=7.4.0
pytest-asyncio>=0.23.0
